pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "frame-annotator"
version = "0.1.0"
description = "A lightweight, configurable web tool for annotating video frames with classification labels."
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    {name = "Omar Choudhry"},
]
keywords = ["annotation", "video", "frames", "labeling", "classification", "research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "flask>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
frame-annotator = "frame_annotator.__main__:main"

[project.urls]
Homepage = "https://github.com/omariosc/frame-annotator"
Repository = "https://github.com/omariosc/frame-annotator"
Issues = "https://github.com/omariosc/frame-annotator/issues"

[tool.setuptools.packages.find]
include = ["frame_annotator*"]
